Add interface for certification
