+++ lib/standards/standards.go
@@ -1,0 +1,5 @@
+package standards
+
+type Base interface {
+
+}
